Synchronized with C++ version for the serialization. Loading of ModifableHDT extracted and separated into two methods: one pass and two passes.
git-svn-id: https://hdt-java.googlecode.com/svn@24 79d02f10-9824-5a6d-8155-d2f4b3853b0f

+++ trunk/hdt-java/src/org/rdfhdt/hdt/compact/integer/VByte3.java
@@ -18,16 +18,16 @@ import java.io.OutputStream;
  *
  */
 public class VByte3 {
-	public static void encode(OutputStream out, int value) throws IOException {
+	public static void encode(OutputStream out, long value) throws IOException {
 		if(value<255) {
-			out.write(value);
+			out.write((int)value);
 		} else {
 			out.write(255);
 			VByte.encode(out, value-255);
 		}
 	}
 	
-	public static int decode(InputStream in) throws IOException {
+	public static long decode(InputStream in) throws IOException {
 		int read = in.read();
 		if(read<255) {
 			return read;
